+++ src/num/ops.c
@@ -1446,8 +1446,8 @@ const struct operator_s* operator_dup_create(const struct operator_s* op, unsign
 	auto ioa = operator_arg_domain(op, a);
 	auto iob = operator_arg_domain(op, b);
 
-	assert(ioa->N == md_calc_blockdim(ioa->N, ioa->dims, ioa->strs, ioa->size));
-	assert(iob->N == md_calc_blockdim(iob->N, iob->dims, iob->strs, iob->size));
+	assert((int)ioa->N == md_calc_blockdim(ioa->N, ioa->dims, ioa->strs, ioa->size));
+	assert((int)iob->N == md_calc_blockdim(iob->N, iob->dims, iob->strs, iob->size));
 
 
 	PTR_ALLOC(struct operator_dup_s, data);
@@ -1590,8 +1590,8 @@ const struct operator_s* operator_link_create(const struct operator_s* op, unsig
 	auto ioi = operator_arg_domain(op, i);
 	auto ioo = operator_arg_domain(op, o);
 
-	assert(ioi->N == md_calc_blockdim(ioi->N, ioi->dims, ioi->strs, ioi->size));
-	assert(ioo->N == md_calc_blockdim(ioo->N, ioo->dims, ioo->strs, ioo->size));
+	assert((int)ioi->N == md_calc_blockdim(ioi->N, ioi->dims, ioi->strs, ioi->size));
+	assert((int)ioo->N == md_calc_blockdim(ioo->N, ioo->dims, ioo->strs, ioo->size));
 
 
 	PTR_ALLOC(struct operator_link_s, data);
@@ -2272,8 +2272,8 @@ const struct operator_s* operator_chainN(unsigned int N, const struct operator_s
 
 			assert(a->domain[0]->N == b->domain[1]->N);
 			assert(md_calc_size(a->domain[0]->N, a->domain[0]->dims) == md_calc_size(b->domain[1]->N, b->domain[1]->dims));
-			assert(a->domain[0]->N == md_calc_blockdim(a->domain[0]->N, a->domain[0]->dims, a->domain[0]->strs, a->domain[0]->size));
-			assert(b->domain[1]->N == md_calc_blockdim(b->domain[1]->N, b->domain[1]->dims, b->domain[1]->strs, b->domain[1]->size));
+			assert((int)a->domain[0]->N == md_calc_blockdim(a->domain[0]->N, a->domain[0]->dims, a->domain[0]->strs, a->domain[0]->size));
+			assert((int)b->domain[1]->N == md_calc_blockdim(b->domain[1]->N, b->domain[1]->dims, b->domain[1]->strs, b->domain[1]->size));
 		}
 	}
